==> Explore/01_csv_to_sql_db.ipynb <==
import pandas as pd
from sqlalchemy import create_engine
from pyprojroot import here
import os

### Set path to input csv files and output SQL database

raw_data_path = here('data/imdb_raw/')
files = os.listdir(raw_data_path)


sql_db_path = f'sqlite:///{here("data/db/sql/imdb.db")}'
engine = create_engine(sql_db_path)

### Read csv files and write to SQL database

for f in files:
    df = pd.read_csv(raw_data_path + f, sep='\t')
    df.replace('\\N', pd.NA, inplace=True)
    df.to_sql(f.rsplit('.', 1)[0], con=engine, if_exists='replace', index=False)

### Write sample data to SQL database

raw_data_path = here('data/imdb_raw/')
sample_db_path = f'sqlite:///{here("data/db/sql/imdb_sample.db")}'

engine_sample = create_engine(sample_db_path)

for f in files:
    df = pd.read_csv(raw_data_path + f, sep='\t').iloc[:20]
    df.replace('\\N', pd.NA, inplace=True)
    df.to_sql(f.rsplit('.', 1)[0], con=engine_sample, if_exists='replace', index=False)
