==> Explore/02_transform_tabular_data.ipynb <==
import os
from dotenv import load_dotenv
from langchain.chains import create_sql_query_chain
from langchain_openai import ChatOpenAI
from langchain_community.utilities import SQLDatabase
from langchain_openai import OpenAIEmbeddings
from pyprojroot import here
import ast
import pandas as pd
load_dotenv()

## Transforming the tabular data
The tabular data is transformed into something that can be embedded into a vector database.
Information about the cast is merged into the movies table, so each row to be embedded contain information about a movie and the cast.

os.environ['OPENAI_API_KEY'] = os.getenv("OPENAI_API_KEY")
llm = ChatOpenAI(model="gpt-3.5-turbo", temperature=0)

sqldb_directory = here("data/db/sql/imdb.db")
db = SQLDatabase.from_uri(f"sqlite:///{sqldb_directory}")

columns = ["tconst", "titleType", "primaryTitle", "originalTitle", "isAdult", "startYear", "endYear", "runtimeMinutes", "genres", "averageRating", "numVotes"]

top_movies_query = """SELECT b.*, r.averageRating, r.numVotes
FROM 'title.basics' b
JOIN 'title.ratings' r ON b.tconst = r.tconst
ORDER BY r.numVotes DESC
LIMIT 100000;
"""

result = db.run(top_movies_query)
parsed_result = ast.literal_eval(result)  

# Load the results into a pandas DataFrame
df_top_movies = pd.DataFrame(parsed_result, columns=columns)

df_top_movies.head()

cast_query = """SELECT p.tconst, p.nconst, p.category, p.job, p.characters
FROM "title.principals" p
WHERE p.tconst IN (
    SELECT tconst FROM "title.ratings" ORDER BY numVotes DESC LIMIT 100000
);
"""

result = db.run(cast_query)
parsed_result = ast.literal_eval(result)

df_cast = pd.DataFrame(parsed_result, columns=["tconst", "nconst", "category", "job", "characters"])

df_cast.head()

names_query = """SELECT * 
FROM "name.basics" 
WHERE nconst IN (
    SELECT DISTINCT nconst 
    FROM "title.principals" 
    WHERE tconst IN (SELECT tconst FROM "title.ratings" ORDER BY numVotes DESC LIMIT 100000)
);
"""

result = db.run(names_query)
parsed_result = ast.literal_eval(result)

df_names = pd.DataFrame(parsed_result, columns=["nconst", "primaryName", "birthYear", "deathYear", "primaryProfession", "knownForTitles"])

df_names.head()

df_top_movies.head()

df_cast_names = df_cast.merge(df_names, on="nconst")

cast_names_grouped = df_cast_names.groupby("tconst").apply(lambda x: "\n".join([f'Name:{row["primaryName"]} {row["birthYear"]}, characters:{row["characters"]}, job:{row["category"]}' for _, row in x.iterrows()]))
cast_names_grouped.name = "cast_info"
df_cast_info = pd.DataFrame(cast_names_grouped.reset_index())

df_cast_info.head()

df_movie_data = df_top_movies.merge(cast_names_grouped, on="tconst")

df_movie_data.to_csv(here("data/imdb_movies_100k.csv"), index=False)
