==> Explore/04_prepare_batch_embedding.ipynb <==
import pandas as pd
import json
from dotenv import load_dotenv
import os
from openai import OpenAI
import tiktoken
from pyprojroot import here

load_dotenv()

os.environ['OPENAI_API_KEY'] = os.getenv("OPENAI_API_KEY")

client = OpenAI() # This automagically uses the key from the environment variable

df = pd.read_csv(here('data/imdb_movies_100k.csv'))

def format_movie_data(movie):
    return f"""
Title: {movie['primaryTitle']} ({movie['startYear']})
Type: {movie['titleType']}
Runtime: {movie['runtimeMinutes']} minutes
Genres: {movie['genres']}
IMDb Rating: {movie['averageRating']} (Votes: {movie['numVotes']})

Cast & Crew:
{movie['cast_info']}
    """

#### Transforming the data the correct format for the OpenAI batch embedding endpoint

batch_data = [{
    "custom_id": f"movie-{i}",
    "method": "POST",
    "url":"/v1/embeddings",
    "body": {
        "model": "text-embedding-3-small",
        "input": format_movie_data(row)
    }
    } for i, row in df.iterrows()]

len(batch_data)

def count_tokens(text, model="text-embedding-3-small"):
    encoding = tiktoken.encoding_for_model(model)
    return len(encoding.encode(text))

### Batch embedding limits
The batch embedding endpoint only allows for 3 million tokens being queued for embedding simultaneously. Therefore, we cannot upload all the batches at the same time. 

We write all the batches to separate files so that they are ready for upload

n_batches = 15 # 41 million token / 15 batches is less than 3 million tokens per batch

n_rows = len(batch_data)
batch_size = n_rows / n_batches + n_rows % n_batches
for i in range(n_batches):
    start_index = i * batch_size
    end_index = min((i + 1) * batch_size, n_rows)
    
    with open(here(f"data/movie_batches/movie_batch_{i+1}.jsonl"), "w") as f:
        for entry in batch_data[start_index:end_index]:
            f.write(json.dumps(entry) + "\n")
