==> Explore/07_comparing_methods.ipynb <==
from langchain_community.utilities import SQLDatabase
from langchain.chains import create_sql_query_chain
from langchain_community.tools.sql_database.tool import QuerySQLDataBaseTool
from langchain_community.agent_toolkits import create_sql_agent
from langchain_core.prompts import PromptTemplate
from langchain_core.output_parsers import StrOutputParser
from langchain_core.runnables import RunnablePassthrough
from langchain_openai import ChatOpenAI, OpenAIEmbeddings
from pyprojroot import here
from operator import itemgetter
import chromadb
import os
from dotenv import load_dotenv
load_dotenv()

### Comparing RAG and SQL Agent for different questions

question1 = "What is the lowest rated movies with more than 1 million votes?"
question2 = "How many movies have Tom Hanks played in?"
question3 = "Name two actors that have played together in more than 4 movies."
question4 = "Name two popular movies about chess."
question5 = "

### SQL

os.environ['OPENAI_API_KEY'] = os.getenv("OPENAI_API_KEY")
llm = ChatOpenAI(model="gpt-3.5-turbo", temperature=0)


sqldb_directory = here("data/db/sql/imdb.db")
db = SQLDatabase.from_uri(f"sqlite:///{sqldb_directory}")

### Simple SQL query writer

system_role = """Given the following user question, corresponding SQL query, and SQL result, answer the user question.\n
    Question: {question}\n
    SQL Query: {query}\n
    SQL Result: {result}\n
    Answer:
    """

execute_query = QuerySQLDataBaseTool(db=db)
write_query = create_sql_query_chain(llm, db)
answer_prompt = PromptTemplate.from_template(system_role)
answer = answer_prompt | llm | StrOutputParser()

chain = (
    RunnablePassthrough.assign(query=write_query)
    .assign(result=itemgetter("query") | execute_query)
    .assign(answer=answer)
)



system_role = """Given the following user question, corresponding SQL query, and SQL result, answer the user question.\n
    Question: {question}\n
    SQL Query: {query}\n
    SQL Result: {result}\n
    Answer:
    """

execute_query = QuerySQLDataBaseTool(db=db)
write_query = create_sql_query_chain(
    llm, db)
answer_prompt = PromptTemplate.from_template(
    system_role)
answer = answer_prompt | llm | StrOutputParser()

chain = (
    RunnablePassthrough.assign(query=write_query)
    .assign(result=itemgetter("query") | execute_query)
    .assign(answer=answer)
)

response1 = chain.invoke({"question": question1})


response1

response = chain.invoke({"question": question2})
response

response = chain.invoke({"question": question3})
response

response = chain.invoke({"question": question4}, config={"verbose": True})
response

### LangChain SQL Agent

agent_executor = create_sql_agent(
    llm, 
    db=db, 
    agent_type="openai-tools", 
    verbose=True, 
    system_message="You are an advanced SQL agent. Your task is to generate accurate and efficient SQL queries to answer user questions based on the database schema and data. The database is the IMDB database with multiple tables and information about movies and tv shows. Do not use the id of actors and movies in the result (e.g. tt00001 and nm000001), instead find the real names."
)

response1_sql = agent_executor.invoke({"input": question1})

response2_sql = agent_executor.invoke({"input": question2})

response3_sql = agent_executor.invoke({"input": question3})

response4_sql = agent_executor.invoke({"input": question4})

### RAG

chroma_client = chromadb.PersistentClient(path=here("data/db/chroma"))
collection_name = "imdb"
collection = chroma_client.get_collection(collection_name)

embedding_client = OpenAIEmbeddings(
    model="text-embedding-3-small",
    # With the `text-embedding-3` class
    # of models, you can specify the size
    # of the embeddings you want returned.
    # dimensions=1024
)

question1_embedding = embedding_client.embed_query(question1)
question1_vector_search = collection.query(
    query_embeddings = question1_embedding,
    n_results=3 #top_k
)
vector_db_results = str(question1_vector_search["documents"])

vector_db_results



system_role = """Given the following user question, and corresponding results from the vector databse, answer the user question.\n
If the vector search results are not related to the question, say "I don't know".\n

Question: {question}\n
Vector search results: {result}\n
Answer:
"""
    
answer_prompt = PromptTemplate.from_template(system_role)
answer = answer_prompt | llm | StrOutputParser()
chain = (
    RunnablePassthrough.assign(embedding=lambda x: embedding_client.embed_query(x["question"]))
    .assign(vector_search=lambda x: collection.query(query_embeddings=x["embedding"], n_results=3))
    .assign(result=lambda x: str(x["vector_search"]["documents"]))
    .assign(answer=answer)
)

response1 = chain.invoke({"question": question1})
print(response1["result"])
print(response1["answer"])

reseponse2 = chain.invoke({"question": question2})
print(reseponse2["result"])
print(reseponse2["answer"])

reseponse3 = chain.invoke({"question": question3})
print(reseponse3["result"])
print(reseponse3["answer"])

reseponse4 = chain.invoke({"question": question4})
print(reseponse4["result"])
print(reseponse4["answer"])
